switched to u8 instead of isize, faster now

+++ src/day04.rs
@@ -3,7 +3,7 @@ pub fn run() -> String {
     let input = parse_input(INPUT);
     format!("Part 1: {}\nPart 2: {}", part1(&input), part2(&input))
 }
-fn parse_input(input: &str) -> Vec<(isize, isize, isize, isize)> {
+fn parse_input(input: &str) -> Vec<(u8, u8, u8, u8)> {
     let mut nums = input
         .split_terminator(&['-', ',', '\n'])
         .map(|s| s.parse().unwrap());
@@ -15,12 +15,12 @@ fn parse_input(input: &str) -> Vec<(isize, isize, isize, isize)> {
     }
     out
 }
-fn part1(input: &Vec<(isize, isize, isize, isize)>) -> usize {
+fn part1(input: &Vec<(u8, u8, u8, u8)>) -> usize {
     input
         .iter()
         .filter(|(a, b, c, d)| (a <= c && b >= d) || (c <= a && d >= b))
         .count()
 }
-fn part2(input: &Vec<(isize, isize, isize, isize)>) -> usize {
+fn part2(input: &Vec<(u8, u8, u8, u8)>) -> usize {
     input.iter().filter(|(a, b, c, d)| a <= d && c <= b).count()
 }
